+++ packages/cubejs-api-gateway/index.js
@@ -382,6 +382,9 @@ class ApiGateway {
       });
       const loadRequestSQLStarted = new Date();
       const normalizedQuery = await this.queryTransformer(normalizeQuery(query), context);
+      if (!normalizedQuery) {
+        throw new Error(`queryTransformer returned null query. Please check your queryTransformer implementation`);
+      }
       const [compilerSqlResult, metaConfigResult] = await Promise.all([
         this.getCompilerApi(context).getSql(coerceForSqlQuery(normalizedQuery, context)),
         this.getCompilerApi(context).metaConfig({ requestId: context.requestId })
@@ -406,17 +409,12 @@ class ApiGateway {
       const response = await this.getAdapterApi({
         ...context, dataSource: sqlQuery.dataSource
       }).executeQuery(toExecute);
-      this.log(context, {
-        type: 'Load Request Success',
-        query,
-        duration: this.duration(requestStarted)
-      });
       const flattenAnnotation = {
         ...annotation.measures,
         ...annotation.dimensions,
         ...annotation.timeDimensions
       };
-      res({
+      const result = {
         query: normalizedQuery,
         data: transformData(aliasToMemberNameMap, flattenAnnotation, response.data, normalizedQuery),
         lastRefreshTime: response.lastRefreshTime && response.lastRefreshTime.toISOString(),
@@ -425,7 +423,13 @@ class ApiGateway {
           usedPreAggregations: response.usedPreAggregations
         }),
         annotation
+      };
+      this.log(context, {
+        type: 'Load Request Success',
+        query,
+        duration: this.duration(requestStarted)
       });
+      res(result);
     } catch (e) {
       this.handleError({
         e, context, query, res, requestStarted