+++ packages/cubejs-server/WebSocketServer.js
@@ -7,10 +7,12 @@ class WebSocketServer {
     options = options || {};
     this.serverCore = serverCore;
     this.processSubscriptionsInterval = options.processSubscriptionsInterval || 5;
+    this.webSocketsBasePath = options.webSocketsBasePath;
   }
 
   initServer(server) {
-    this.wsServer = new WebSocket.Server({ server });
+    this.wsServer = this.webSocketsBasePath ?
+      new WebSocket.Server({ server, path: this.webSocketsBasePath }) : new WebSocket.Server({ server });
 
     const connectionIdToSocket = {};